refactor(Input): combine function into one retry function and apply
